# QLS: check `NoThunks` only at the end of an action sequence