@jorisdral 2025-01-20
> Nice. And does it improve test performance at all? How much?

I've run the tests with assertions enabled/disabled and before/after the PR.

Assertions enabled: the performance improves significantly for `RealImpl_RealFS_IO` and `RealImpl_MockFS_IO`. These are also the only two scenarios that should be affected.

|   | `ModelIOImpl`  | `RealImpl_RealFS_IO`  | `RealImpl_MockFS_IO` | `RealImpl_MockFS_IOSim`  |
|---|---|---|---|---|
| After | 4.91s | 11.26s | 21.38s | 27.57s |
| Before | 5.27s | 43.07s | 51.58s | 28.95s |

Assertions disabled: no difference, except for the runs with the mocked FS. I don't think I changed anything meaningful that would explain the slowdown, so I'm just assuming it's noise

|   | `ModelIOImpl`  | `RealImpl_RealFS_IO`  | `RealImpl_MockFS_IO` | `RealImpl_MockFS_IOSim`  |
|---|---|---|---|---|
| After | 4.20s | 7.74s | 22.44s | 34.09s |
| Before | 4.16s | 7.90s | 20.94s | 31.89s |